Refactor bonome wizard into modular phases
